\name{readSimFFPE}
\alias{readSimFFPE}
\title{
Simulate noisy NGS reads of FFPE samples for whole genome / several chromosomes
/ large regions
}
\description{
NGS read from FFPE samples contain numerous artificial chimeric reads. This
function simulates these artifacts as well as normal reads for FFPE samples
on whole genome, or several chromosomes, or large regions.
}
\usage{
readSimFFPE(sourceSeq, reference, PhredScoreProfile, outFile, coverage, readLen,
meanInsertLen, sdInsertLen, enzymeCut = FALSE, covRatioFFPE = 0.08,
localMatchRatio = 0.1, windowLen = 10000, matchWinLen = 10000, meanSeedLen = 5,
sdSeedLen = 3, revChimericProb = 0.8, spikeWidth = 1500, mutationRate = 0.005,
noiseRate = 0.0015, highNoiseRate = 0.08, highNoiseProb = 0.015,
pairedEnd = TRUE, prefix = "SimFFPE", threads = 1, localChimeric = TRUE,
distantChimeric = TRUE, normalReads = TRUE, overWrite = FALSE)
}
\arguments{
  \item{sourceSeq}{
A DNAStringSet object of the DNA sequences to be simulated. It can cover the
whole reference genome or just the selected chromosomes or chromosome regions.
}
  \item{reference}{
A DNAStringSet object of the whole reference genome.
}
  \item{PhredScoreProfile}{
A matrix representing the positional Phred score proportion. Each row of the 
matrix represents a position in the read (from begin to end), and each column 
the Phred quality score of base-calling error probabilities. The profile can be 
calculated from BAM file using the \code{\link{estimateSimParam}} function.
}
  \item{outFile}{
Output file path for the FASTQ file with simulated reads. Please include the
name of the output file, e.g. "/tmp/sim1".
}
  \item{coverage}{
Coverage of the simulated file.
}
  \item{readLen}{
Read length of the simulated reads.
}
  \item{meanInsertLen}{
Mean insert length for the simulation (normally distributed).
}
  \item{sdInsertLen}{
Standard deviation of the insert length for simulation (normally distributed).
}
  \item{enzymeCut}{
Simulate enzymatic fragmentation if it is set to true, otherwise simulate random
fragmentation.
}
  \item{covRatioFFPE}{
Proportion of reads generated from artifact chimeric sequences (the number of
reads from chimeric sequences divided by the number of all reads including
chimeric reads and normal reads). Range: 0 to 1.
}
  \item{localMatchRatio}{
Proportion of chimeric reads that are formed by the combination of two nearby
DNA fragments (the number of chimeric reads of nearby DNA combination divided by
the number of all chimeric reads including chimeric reads derived from nearby or
distant DNA combination). Range: 0 to 1.
}
  \item{windowLen}{
The local window length used in the simulation of the combination of nearby DNA
fragments. To simulate nearby DNA fragments combination, the input DNA sequences
are divided into small regions with same window length, and the short reverse
complementary sequences will be searched within the same region. The larger the
window, the higher the probability of successfully finding a short reverse
complementary match, but the greater the allowed maximum distance between two
nearby DNA fragments for combination and the longer the runtime. Suggested
range: 5000-20000. Unit: base pair (bp).
}
  \item{matchWinLen}{
The target window length used in the simulation of the combination of distant
DNA fragments. To simulate distant DNA fragments combination, the reference DNA
sequences are divided into small regions with same window length, and the short
reverse complementary sequence will be searched in a random region. The larger
the window, the higher the probability of successfully finding a short reverse
complementary match, but the longer the runtime. Suggested range: 5000-20000.
Unit: base pair (bp).
}
  \item{meanSeedLen}{
Mean length of short seed sequences used to search for reverse complementary
matches in chimeric sequences simulation (normally distributed length, with a
built-in minimum of 2 and maximum of 20). The longer it is, the lower the
probability of successfully finding a short reverse complementary match.
Suggested range: 4-8. Unit: base pair (bp).
}
  \item{sdSeedLen}{
Standard deviation of short seed sequences used to search for reverse
complementary matches in chimeric sequences simulation (normally distributed
length, with a built-in minimum of 2 and maximum of 20). Suggested range: 2-5.
}
  \item{revChimericProb}{
The probability of local-chimeric sequences to be generated through the
combination of a DNA fragment and its duplicate in reverse direction (the number
of local-chimeric sequences generated by the matches on the same strand in
reverse direction divided by the number of all local-chimeric sequences
generated by the matches on the same or complementary strand). For paired end
sequencing simulation, the larger the probability, the greater the proportion of
improper pair-orientated reads with LL / RR pair orientation, and the smaller
with RL pair orientation. For distant-chimeric sequences, this probability is
with a built-in value of 0.5. Range: 0 to 1. Suggested range: 0.7-0.9 based on
the real FFPE samples to simulate.
}
  \item{spikeWidth}{
The width of chimeric read spike used in the simulation of the combination of
distant DNA fragments. In real FFPE samples, the chimeric reads formed by
distant DNA combination are unevenly distributed along the chromosome. Some
regions are enriched in these distant-chimeric reads and some are not. The
length of these regions are of similar scale and therefore a defined width is
used for simulation. Suggested range: 1500-2000. Unit: base pair (bp).
}
  \item{mutationRate}{
Mutation rate for each base in chimeric sequences. In the whole chimeric
sequence formation process, some biological-level errors might occur and lead to
mutations. The probability is set to be even for all four basic type of
nucleotides. Range: 0-0.75.
}
  \item{noiseRate}{
Noise rate for each base in reads. This is used for sequencing-level errors.
The probability is set to be even for all four basic type of nucleotides.
Range: 0-0.75.
}
  \item{highNoiseRate}{
A second noise rate for each base in reads. This is used to simulate the
sequencing-level errors. The probability is set to be even for all four basic
type of nucleotides. In some real sequencing data, some reads are much more
noisy than others and this parameter can be used to simulate the noises on these
reads. Range: 0-0.75.
}
  \item{highNoiseProb}{
Probability of reads to be simulated with highNoiseRate other than noiseRate.
Range: 0-1.
}
  \item{pairedEnd}{
Simulate paired end sequencing when set to true.
}
  \item{prefix}{
Prefix for read names. When reads from different runs of simulation have to be
merged, please make sure that they have different prefixes.
}
  \item{threads}{
Number of threads used. Multi-threading can speed up the process.
}
  \item{localChimeric}{
Generate reads from local-chimeric sequences if it is set to true. If it is set
to false, skip this process.
}
  \item{distantChimeric}{
Generate reads from distant-chimeric sequences if it is set to true. If it is
set to false, skip this process.
}
  \item{normalReads}{
Generate reads from normal sequences if it is set to true. If it is set to
false, skip this process.
}
  \item{overWrite}{
Overwrite the file if file with same output path exists and it is set to true.
If file with same output path exists and it is set to false, reads will be
appended to the existing file.
}
}
\details{
The NGS (Next-Generation Sequencing) reads from FFPE
(Formalin-Fixed Paraffin-Embedded) samples contain numerous artificial chimeric
reads. These reads are derived from the combination of two single-stranded DNA
fragments with short reverse complementary sequences. This function simulates
these artifacts as well as normal reads for FFPE samples on whole genome /
several chromosomes / large regions. The DNA fragments to be combined can be
derived from nearby regions or distant regions. In the output fastq file these
reads are distinguished by "localChimeric", "distantChimeric" and "Normal" in
their names. The parameter PhredScoreProfile can be derived by the function
\code{\link{estimateSimParam}}. To simulate whole exome sequencing (WES) or
targeted sequencing, please use the function \code{\link{targetReadSimFFPE}}.
}
\value{
NULL. Reads will be written to the output FASTQ file.
}
\author{
Lanying Wei <lanying.wei@uni-muenster.de>
}
\note{
When fine-tuning is needed, simulate reads from certain areas / chromosomes
instead of the entire genome to save the runtime. Please check the package
vignette for the guidance of fine-tuning.
}
\seealso{
\code{\link{SimFFPE}}, \code{\link{estimateSimParam}},
\code{\link{targetReadSimFFPE}}
}
\examples{
PhredScoreProfilePath <- system.file("extdata", "PhredScoreProfile2.txt",
                                      package = "SimFFPE")
PhredScoreProfile <- as.matrix(read.table(PhredScoreProfilePath, skip = 1))
colnames(PhredScoreProfile) <- read.table(PhredScoreProfilePath, 
                                          nrows = 1, colClasses = "character")

referencePath <- system.file("extdata", "example.fasta", package = "SimFFPE")
reference <- readDNAStringSet(referencePath)

## Simulate reads of the first three sequences of reference genome

sourceSeq <- reference[1:3]
outFile1 <- paste0(tempdir(), "/sim1")
readSimFFPE(sourceSeq, reference, PhredScoreProfile, outFile1, overWrite = TRUE,
            coverage = 80, readLen = 150, meanInsertLen = 250, sdInsertLen = 80,
            enzymeCut = TRUE, threads = 2)


## Simulate reads of defined regions on the first two sequences of reference
## genome

sourceSeq2 <- DNAStringSet(lapply(reference[1:2], function(x) x[1:10000]))
outFile2 <- paste0(tempdir(), "/sim2")
readSimFFPE(sourceSeq2, reference, PhredScoreProfile, outFile2,
            overWrite = TRUE, coverage = 80, readLen = 150, meanInsertLen = 250,
            sdInsertLen = 80, enzymeCut = TRUE, threads = 1)


## Simulate reads of defined regions on the second and the third sequence of 
## reference genome and merge them with existing reads (a different prefix is 
## needed)

sourceSeq3 <- DNAStringSet(lapply(reference[2:3], function(x) x[1:10000]))
readSimFFPE(sourceSeq3, reference, PhredScoreProfile, outFile2,
            overWrite = FALSE, prefix = "simFFPE2", coverage = 80,
            readLen = 150, meanInsertLen = 250, sdInsertLen = 80,
            enzymeCut = TRUE, threads = 1)
}

